test: move setup function to new file
